add refactoring of house class
