+++ document/javadoc/org.team4/org/team4/house/House.html
@@ -89,7 +89,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="contentContainer">
 <div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">org.team4.house.House</div>
+<div class="inheritance">org.team4.house.components.House</div>
 </div>
 <section class="description">
 <hr>
