+++ src/main/java/org/team4/common/observer/SimulatorObserver.java
@@ -2,16 +2,22 @@ package org.team4.common.observer;
 
 import org.team4.common.Settings;
 import org.team4.common.logger.Logger;
-import org.team4.house.House;
+import org.team4.house.services.LightService;
+import org.team4.user.UserService;
+
 import java.util.Observable;
 import java.util.Observer;
 
 @Deprecated
 public class SimulatorObserver implements Observer {
 
+    private LightService lightService;
+    private UserService userService;
+
     @Deprecated
     public SimulatorObserver(Observable o) {
         o.addObserver(this);
+        lightService = new LightService();
     }
 
     @Override
@@ -31,7 +37,7 @@ public class SimulatorObserver implements Observer {
      * Performs the away mode actions
      */
     public void performAwayModeChecks() {
-        if(House.userInHouse()) {
+        if(userService.userInHouse()) {
             Logger.warning("Movement detected inside the house.");
             Settings.startTimeBeforeAlerting = true;
         }
@@ -41,7 +47,7 @@ public class SimulatorObserver implements Observer {
      * Performs the auto mode action
      */
     public void performAutoModeCheck() {
-        House.turnOnAllLightsWithUsers();
+        lightService.turnOnAllLightsWithUsers();
 
     }
