+++ src/test/java/org/team4/delivery2Test/u10ChooseLightInAwayModeTest.java
@@ -2,13 +2,13 @@ package org.team4.delivery2Test;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.team4.house.House;
+import org.team4.house.services.LightService;
 
 public class u10ChooseLightInAwayModeTest {
     @Test
     public void testAddtoAwayModeTest() {
         try {
-            House.toggleLightsAway("aa");
+            new LightService().toggleLightsAway("aa");
         }
         catch (Exception e) {
             Assert.assertTrue(true);
