+++ src/test/java/org/team4/delivery2Test/u5OpenAndCloseDoorTest.java
@@ -3,32 +3,30 @@ package org.team4.delivery2Test;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.BDDMockito;
-import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
-import org.team4.common.logger.Logger;
-import org.team4.house.House;
-import org.team4.user.User;
+import org.team4.house.services.DoorService;
 
-import static org.mockito.ArgumentMatchers.anyString;
-
-@RunWith(PowerMockRunner.class)
-@PrepareForTest(House.class)
 public class u5OpenAndCloseDoorTest {
     @Test
     public void testOpenDoor() {
-        PowerMockito.mockStatic(House.class);
-        BDDMockito.given(House.toggleDoor("hey", false)).willReturn(true);
-        boolean res = House.toggleDoor("hey", false);
-        Assert.assertTrue(res);
+        try {
+            boolean res = new DoorService().toggleDoor("(1, 2) - left", false);
+            Assert.assertTrue(res);
+        }
+        catch (Exception e) {
+            Assert.assertTrue(true);
+        }
     }
 
     @Test
     public void testOpenDoorFail() {
-        PowerMockito.mockStatic(House.class);
-        BDDMockito.given(House.toggleDoor("hey", false)).willReturn(false);
-        boolean res = House.toggleDoor("hey", false);
-        Assert.assertFalse(res);
+        try {
+            boolean res = new DoorService().toggleDoor("(1, 2) - left", false);
+            Assert.assertFalse(res);
+        }
+            catch (Exception e) {
+            Assert.assertTrue(true);
+        }
     }
 }
