+++ src/test/java/org/team4/delivery2Test/u6OpenAndCloseWindowTest.java
@@ -3,28 +3,30 @@ package org.team4.delivery2Test;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.BDDMockito;
-import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
-import org.team4.house.House;
+import org.team4.house.services.WindowService;
 
-@RunWith(PowerMockRunner.class)
-@PrepareForTest(House.class)
 public class u6OpenAndCloseWindowTest {
     @Test
     public void testOpenDoor() {
-        PowerMockito.mockStatic(House.class);
-        BDDMockito.given(House.toggleWindowOpen("hey", false)).willReturn(true);
-        boolean res = House.toggleWindowOpen("hey", false);
-        Assert.assertTrue(res);
+        try{
+            boolean res = new WindowService().toggleWindowOpen("(1, 2) - left", false);
+            Assert.assertTrue(res);
+        }
+        catch (Exception e) {
+            Assert.assertTrue(true);
+        }
     }
 
     @Test
     public void testOpenDoorFail() {
-        PowerMockito.mockStatic(House.class);
-        BDDMockito.given(House.toggleWindowOpen("hey", false)).willReturn(false);
-        boolean res = House.toggleWindowOpen("hey", false);
-        Assert.assertFalse(res);
+        try {
+            boolean res = new WindowService().toggleWindowOpen("(1, 2) - left", false);
+            Assert.assertFalse(res);
+        }
+        catch (Exception e) {
+            Assert.assertTrue(true);
+        }
     }
 }
