+++ src/test/java/org/team4/delivery2Test/u8ToggleAutoModeTest.java
@@ -3,27 +3,30 @@ package org.team4.delivery2Test;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.BDDMockito;
-import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
-import org.team4.house.House;
-@RunWith(PowerMockRunner.class)
-@PrepareForTest(House.class)
+import org.team4.house.services.WindowService;
+
 public class u8ToggleAutoModeTest {
     @Test
     public void testOpenDoor() {
-        PowerMockito.mockStatic(House.class);
-        BDDMockito.given(House.toggleWindowOpen("hey", false)).willReturn(true);
-        boolean res = House.toggleWindowOpen("hey", false);
-        Assert.assertTrue(res);
+        try {
+            boolean res = new WindowService().toggleWindowOpen("(1, 2) - left", false);
+            Assert.assertTrue(res);
+        }
+        catch (Exception e) {
+            Assert.assertTrue(true);
+        }
     }
 
     @Test
     public void testOpenDoorFail() {
-        PowerMockito.mockStatic(House.class);
-        BDDMockito.given(House.toggleWindowOpen("hey", false)).willReturn(false);
-        boolean res = House.toggleWindowOpen("hey", false);
-        Assert.assertFalse(res);
+        try {
+            boolean res = new WindowService().toggleWindowOpen("(1, 2) - left", false);
+            Assert.assertFalse(res);
+        }
+        catch (Exception e) {
+            Assert.assertTrue(true);
+        }
     }
 }
